move isFormReadonly to EtoApiUtils.ts
