+++ app/components/eto/registration/pages/EtoEquityTokenInfo.tsx
@@ -9,7 +9,7 @@ import {
   EtoEquityTokenInfoType,
   TPartialEtoSpecData,
 } from "../../../../lib/api/eto/EtoApi.interfaces";
-import { etoFromIsReadonly } from "../../../../lib/api/eto/EtoUtils";
+import { etoFromIsReadonly } from "../../../../lib/api/eto/EtoApiUtils";
 import { actions } from "../../../../modules/actions";
 import { selectIssuerEto, selectIssuerEtoState } from "../../../../modules/eto-flow/selectors";
 import { EEtoFormTypes } from "../../../../modules/eto-flow/types";
